% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_stocks_data.R
\name{carbon_content}
\alias{carbon_content}
\title{Save the data of the amount of carbon stored in soils (1), in the biomass outside
of forests (2), and in forests (3) from the ALDO Excel tool.
The data is stored in the "data" directory}
\usage{
carbon_content(path_to_aldo)
}
\arguments{
\item{path_to_aldo}{}
}
\value{
None
}
\description{
Save the data of the amount of carbon stored in soils (1), in the biomass outside
of forests (2), and in forests (3) from the ALDO Excel tool.
The data is stored in the "data" directory
}
