% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_aldo_data.R
\name{carbon_content_biomass_wo_forests}
\alias{carbon_content_biomass_wo_forests}
\title{Extracts the data for the amount of carbon stored in biomass outside of
forests from the Aldo Excel tool.}
\usage{
carbon_content_biomass_wo_forests(path_to_aldo)
}
\arguments{
\item{path_to_aldo}{}
}
\value{
A data.table object.
}
\description{
Extracts the data for the amount of carbon stored in biomass outside of
forests from the Aldo Excel tool.
}
