% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_flows_data.R
\name{carbon_flows}
\alias{carbon_flows}
\title{Save the data of carbon flows from the aldo tool
The data is stored in the "data -> aldo" directory}
\usage{
carbon_flows(path_to_aldo)
}
\value{
None
}
\description{
Save the data of carbon flows from the aldo tool
The data is stored in the "data -> aldo" directory
}
