% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_flows.R
\name{carbon_flows_soil}
\alias{carbon_flows_soil}
\title{Extracts the data for carbon flows in soils that occur because
of land use changes from the Aldo Excel tool.}
\usage{
carbon_flows_soil(path_to_aldo)
}
\arguments{
\item{path_to_aldo}{}
}
\value{
A data.table object.
}
\description{
Extracts the data for carbon flows in soils that occur because
of land use changes from the Aldo Excel tool.
}
