% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctdat.R
\docType{data}
\name{ctdat}
\alias{ctdat}
\title{Example CTD Data from a Valeport SWiFT Device.}
\format{
An object of class "list" containing multiple profiles as recorded by
a Valeport SWiFT CTD device.
}
\usage{
data(ctdat)
}
\description{
Example CTD Data from a Valeport SWiFT Device.
}
\examples{
## Load in example data and set it up in the temporary directory:
setup_example()

## Read the example casts:
casts <- read_vp2(directory = tempdir(), type = "Chlorophyll", ID = 12345)

}
\keyword{datasets}
