% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_bd_foret.R
\name{download_bd_foret}
\alias{download_bd_foret}
\title{Download bd foret layer on the territory, for the selected years,
using 'happign' package
The function returns nothing but store gpkg files
in the "data->arep" folder}
\usage{
download_bd_foret(shape)
}
\arguments{
\item{shape}{shape of the territory}
}
\value{
None
}
\description{
Download bd foret layer on the territory, for the selected years,
using 'happign' package
The function returns nothing but store gpkg files
in the "data->arep" folder
}
