% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gpkg_files.R
\name{download_gpkg_files}
\alias{download_gpkg_files}
\title{Download the territory chosen with the data from Corine Land Cover
of the chosen years, using 'happign' package
The function returns nothing but store the gpkg file
in the "data->arep" folder}
\usage{
download_gpkg_files(years, data_path, bd_foret = T)
}
\arguments{
\item{years}{years to download}

\item{data_path}{path to where the data is stored}

\item{bd_foret}{boolean to download bd_foret}
}
\description{
Download the territory chosen with the data from Corine Land Cover
of the chosen years, using 'happign' package
The function returns nothing but store the gpkg file
in the "data->arep" folder
}
