% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_impermability_layers.R
\name{download_impermability_layer}
\alias{download_impermability_layer}
\title{Download the impermability layer}
\usage{
download_impermability_layer(shape, data_path)
}
\arguments{
\item{shape}{shape of the territory}

\item{data_path}{path to where the data is stored}
}
\description{
Download the impermability layer on the territory, for the selected years,
using 'happign' package
}
