% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_impermability_layers.R
\name{download_impermability_layers}
\alias{download_impermability_layers}
\title{Download impermability layers on the territory, for the selected years,
using 'happign' package
The function returns nothing but store raster files
in the "data->copernicus" folder}
\usage{
download_impermability_layers(shape, data_path, years = c("12", "15"))
}
\arguments{
\item{shape}{shape of the territory}

\item{data_path}{path to where the data is stored}

\item{years}{years to download}
}
\description{
Download impermability layers on the territory, for the selected years,
using 'happign' package
The function returns nothing but store raster files
in the "data->copernicus" folder
}
