% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_territory_files.R
\name{download_territory_files}
\alias{download_territory_files}
\title{Download territory files}
\usage{
download_territory_files(
  communes_fr = list(),
  epcis_fr = list(),
  departments_fr = list(),
  regions_fr = list(),
  years,
  data_path
)
}
\arguments{
\item{years}{years to download}

\item{data_path}{path to where the data is stored}

\item{bd_foret}{boolean to download bd_foret}
}
\description{
Download the territory chosen with the data from Corine Land Cover
of the chosen years, using 'happign' package
The function returns nothing but store the gpkg file
in the "data->arep" folder
}
