% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles.R
\name{draw_plot}
\alias{draw_plot}
\title{Plot Profile}
\usage{
draw_plot(var, data)
}
\arguments{
\item{var}{Name of the variable to be plotted.}

\item{data}{Data containing the variable to be plotted.}
}
\value{
ggplot of the selected variable.
}
\description{
Plots the profile for a selected variable from the CTD data-set. The plot
consists of all data points, with those labeled as incongruent colored red.
The moving average for a window length of 5 points is shown as a blue line.
}
\examples{
## Load in example data and set it up in the temporary directory:
setup_example()

## Read the example casts:
casts <- read_vp2(directory = tempdir(), type = "Chlorophyll", ID = 12345)

## Label incongruents in the imported data:
output <- label_incongruents(
  df1 = casts, W = 0.6, alpha = 0.001,
  type = "Chlorophyll"
)

## Select the data for one file:
selected_filename <- unique(output$filename)[1]
single_file <- output[which(output$filename == selected_filename), ]

## Plot a single profile:
draw_plot(var = "temp", data = single_file)

## To clear the temporary directory after using 'setup_example()':
unlink(paste0(tempdir(), "\\\\", list.files(tempdir(), pattern = ".vp2")))

}
