% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_aldo_carbon_flows.R
\name{extract_aldo_carbon_flows}
\alias{extract_aldo_carbon_flows}
\title{Extracts the data for carbon flows from the Aldo Excel tool.}
\usage{
extract_aldo_carbon_flows(path_to_aldo)
}
\arguments{
\item{path_to_aldo}{}
}
\value{
None
}
\description{
Extracts the data for carbon flows from the Aldo Excel tool.
}
