% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_aldo_carbon_stocks.R
\name{extract_aldo_carbon_stocks}
\alias{extract_aldo_carbon_stocks}
\title{Extract Aldo carbon stocks}
\usage{
extract_aldo_carbon_stocks()
}
\arguments{
\item{path_to_aldo}{the path to the excel tool Aldo}
}
\value{
None
}
\description{
Extracts the data for the amount of carbon stored in soils from the Aldo
Excel tool.
}
