% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apigouv_insee_codes.R
\name{get_apigouv_insee_codes}
\alias{get_apigouv_insee_codes}
\title{Apigouv insee codes}
\usage{
get_apigouv_insee_codes(
    communes_fr,
    epcis_fr,
    departments_fr,
    regions_fr)
}
\arguments{
\item{communes_fr}{A \code{character} or a list of \code{character} of french
cities insee codes}

\item{epcis_fr}{A \code{character} or a list of \code{character} of french
epcis insee codes}

\item{departments_fr}{A \code{character} or a list of \code{character} of french
department insee codes}

\item{regions_fr}{A \code{character} or a list of \code{character} of french
regions insee codes}
}
\value{
\code{get_apigouv_insee_codes} returns a list of cities and departments insee codes
}
\description{
Retrieve insee codes from either epcis codes of departement codes
Codes of the different french territories
(communes, EPCI, départements, régions) are  referenced here
\href{https://www.insee.fr/fr/information/6051727}{INSEE's codes}

The function returns a list of the INSEE codes of the cities
located in the selected territory".
}
