% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_carbon_flows.R
\name{get_carbon_flows}
\alias{get_carbon_flows}
\title{Get carbon flows}
\usage{
get_carbon_flows(year_from, year_to, data_path)
}
\arguments{
\item{year_from}{the CLC reference year (1990, 2000, 2006, 2012, 2018)}

\item{year_to}{the year to be compared to the reference year,
must be superior to reference year}

\item{data_path}{path to where the data is stored}
}
\value{
a sf object with the carbon flows of the region
}
\description{
Retrieve the carbon flows between two years, of a chosen region
}
