% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_carbon_flows_balance.R
\name{get_carbon_flows_balance}
\alias{get_carbon_flows_balance}
\title{Get carbon flows balance}
\usage{
get_carbon_flows_balance(flows)
}
\arguments{
\item{flows}{sf object returned by \code{get_carbon_flows()} function}
}
\value{
a data.table with the flows in ktCO2/year
}
\description{
Compute the carbon flows balance between the two years, of the chosen territory.

There are four flows categories :
\itemize{
\item carbon flows from land use changes (ktCO2e/an)
\item carbon flows from forest (ktCO2e/an)
\item carbon flows from harvested wood (ktCO2e/an)
\item total carbon flows (ktCO2e/an)
}
}
