% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_carbon_stocks.R
\name{get_carbon_stocks}
\alias{get_carbon_stocks}
\title{Get carbon stocks}
\usage{
get_carbon_stocks(year)
}
\arguments{
\item{year}{the reference year to be studied (1990, 2000, 2006, 2012, 2018)}

\item{data_path}{path to where the data is stored}
}
\value{
a sf object with the carbon stored in every polygon
}
\description{
\code{get_carbon_stocks} calculates the carbon stocks of the selected
territory for the selected year.
All stocks are in tCO2e/ha.
}
\examples{
# Retrieve carbon stocks in 2018

dt <- get_carbon_stocks("2018", data_path)

}
