% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_carbon_stocks.R
\name{get_carbon_storage}
\alias{get_carbon_storage}
\title{Retrieve carbon stocks of the chosen region}
\usage{
get_carbon_storage(year, count = F)
}
\value{
a sf object with the carbon stored in every polygon
}
\description{
Retrieve carbon stocks of the chosen region
}
