% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_carbon_flows.R
\name{get_forest_flows}
\alias{get_forest_flows}
\title{Retrieve the carbon flows in the forest,
using the BD foret 2}
\usage{
get_forest_flows()
}
\value{
forest flows in tC.an-1
}
\description{
Retrieve the carbon flows in the forest,
using the BD foret 2
}
