% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_carbon_flows.R
\name{get_total_carbon_flows}
\alias{get_total_carbon_flows}
\title{Compute total carbon flows between two years, of a chosen region.
Four flows categories : biomass from land use changes, soil from land
use changes, total forest flows (photosynthesis), total area flows}
\usage{
get_total_carbon_flows(flows, bd = T)
}
\arguments{
\item{flows}{st object returned by 'get_carbon_flows' funciton}

\item{bd}{boolean to use bd foret or not}
}
\value{
a data.table with the flows in ktCO2/year
}
\description{
Compute total carbon flows between two years, of a chosen region.
Four flows categories : biomass from land use changes, soil from land
use changes, total forest flows (photosynthesis), total area flows
}
