% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_incongruents.R
\name{label_incongruents}
\alias{label_incongruents}
\title{Label Incongruents}
\usage{
label_incongruents(
  df1,
  W,
  alpha,
  iterations = 1,
  type = c("Chlorophyll", "Turbidity"),
  method = c("t.student", "max.residual", "chisq", "dixon", "grubbs")
)
}
\arguments{
\item{df1}{Data frame containing data from .vp2 files as obtained from
\code{\link[=read_vp2]{read_vp2()}}.}

\item{W}{Window width against which each data-point is compared along the CTD
depth profile. Units should match those expressed in the "depth" column of
the input data.}

\item{alpha}{Level of significance used to test the incongruence of
data-points. Larger alpha-values make the function more sensitive in the
identification of incongruents.}

\item{iterations}{Number of iterations for applying the incongruence labeling
test. Points labeled "incongruent" are ignored in subsequent iterations.}

\item{type}{A character string. Either "Chlorophyll" or "Turbidity" for the
appropriate CTD measurement type.}

\item{method}{Method to be used to determine whether a point is an outlier.
Should be one of "t.student", "max.residual", "chisq", "dixon", or "grubbs".}
}
\value{
List. Contains two data-frames. "labeled_data" contains the input
data-frame with columns labeling the incongruence of each point for each
variable; the "warning" column indicating whether a certain measurement depth
has incongruent values for more than two variables.
}
\description{
Creates additional columns to a data-frame in the format obtained from
\code{\link[=read_vp2]{read_vp2()}}, indicating for each variable weather a data point is
incongruent, and the associated p-value for the observation. Additionally it
creates a "warnings" column indicating whether three or more variables for a
given measurement depth show incongruent values.
}
\details{
Incongruence is determined with either a t-test or one of the methods
available from the package "outliers": with a Chi-squared test, Dixon test,
Grubbs test, or by determining if a point has the highest residual from the
mean of the window specified.
}
\examples{
## Load in example data and set it up in the temporary directory:
setup_example()

## Read the example casts:
casts <- read_vp2(directory = tempdir(), type = "Chlorophyll", ID = 12345)
casts <- casts[which(casts$filename \%in\% unique(casts$filename)[1:2]),]

## Label incongruents in the imported data:
output <- label_incongruents(
  df1 = casts, W = 0.6, alpha = 0.0001,
  type = "Chlorophyll",
  method = "t.student"
)

## To clear the temporary directory after using 'setup_example()':
unlink(paste0(tempdir(), "\\\\", list.files(tempdir(), pattern = ".vp2")))

}
