% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ign_data.R
\name{load_ign_aldo_data}
\alias{load_ign_aldo_data}
\title{Load ign aldo data}
\usage{
load_ign_aldo_data()
}
\value{
None
}
\description{
Check if Write a \code{cities.gpkg} files with data on french  cities.
}
