% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_and_plots.R
\name{map_carbon_flows}
\alias{map_carbon_flows}
\title{Plot carbon flows of a chosen region.}
\usage{
map_carbon_flows(flows, data_path)
}
\arguments{
\item{flows}{st object returned by 'get_carbon_flows' funciton}

\item{data_path}{path to where the data is stored}
}
\value{
map of carbon flows
}
\description{
Plot carbon flows of a chosen region.
}
