% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_and_plots.R
\name{map_carbon_storage}
\alias{map_carbon_storage}
\title{Map the carbon storage on the chosen region.}
\usage{
map_carbon_storage(dt, data_path)
}
\arguments{
\item{data_path}{path to where the data is stored}
}
\value{
a ggplot with the map
}
\description{
Map the carbon storage on the chosen region.
}
