% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moving_fns.R
\name{moving_fn}
\alias{moving_fn}
\title{Moving Function}
\usage{
moving_fn(x, w, fun, side, ...)
}
\arguments{
\item{x}{Vector with numeric data.}

\item{w}{Window length.}

\item{fun}{Function to apply.}

\item{side}{Side to take, (c)entre, (l)eft or (r)ight.}

\item{...}{Parameters passed on to 'fun'.}
}
\value{
Vector containing the function outputs for each window along the
vector x.
}
\description{
Applies a specified function to subsets of a vector sequentially to compute a
'moving window' along the the vector.
The window is defined by each element as either the subsequent (right),
preceding (left), or neighboring (center) elements.
}
\examples{
my_data <- rnorm(50, mean = 30, sd = 5)
moving_averages <- moving_fn(my_data, w = 5, fun = mean, side = "center")

}
