% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_carbon_flows.R
\name{plot_carbon_flows}
\alias{plot_carbon_flows}
\title{Plot carbon flows}
\usage{
plot_carbon_flows(flows, data_path)
}
\arguments{
\item{flows}{sf object returned by \code{get_carbon_flows()} function}

\item{data_path}{path to where the data is stored}
}
\value{
map of carbon flows
}
\description{
Plot carbon flows of a chosen region.
}
