% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_carbon_storage.R
\name{plot_carbon_stocks}
\alias{plot_carbon_stocks}
\title{Plot carbon stocks}
\usage{
plot_carbon_stocks(stocks, data_path)
}
\arguments{
\item{stocks}{sf object returned by \code{get_carbon_stocks()} function}

\item{data_path}{path to where the data is stored}
}
\value{
a ggplot with the map
}
\description{
Plot the carbon stocks on the chosen territory
}
