% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_carbon_storage.R
\name{plot_carbon_storage}
\alias{plot_carbon_storage}
\title{Plot carbon storage}
\usage{
plot_carbon_storage(year, data_path)
}
\arguments{
\item{year}{the reference year to be studied (1990, 2000, 2006, 2012, 2018)}

\item{data_path}{path to where the data is stored}
}
\value{
a ggplot with the map
}
\description{
Plot the carbon storage on the chosen territory
}
