% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest_flows.R
\name{plot_forest_flows}
\alias{plot_forest_flows}
\title{Plot forest flows}
\usage{
plot_forest_flows(data_path)
}
\arguments{
\item{data_path}{path to where the data is stored}
}
\value{
map of carbon flows by the forest
}
\description{
Plot carbon flows in forest of the chosen territory.
}
