% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_land_use_changes.R
\name{plot_land_use_changes}
\alias{plot_land_use_changes}
\title{Plot land use changes}
\usage{
plot_land_use_changes(year_from, year_to, data_path)
}
\arguments{
\item{year_from}{the CLC reference year (1990, 2000, 2006, 2012, 2018)}

\item{year_to}{the year to be compared to the reference year,
must be superior to reference year}

\item{data_path}{path to where the data is stored}
}
\value{
The sankey diagram
}
\description{
Plot land use changes between two year.
Results are presented in a sankey diagram, with the reference year on the left,
and the year to be compared to on the right.
Note that only soils that have changed occupation between the two years are considered.
}
\examples{
# Plot land use changes between 1990 and 2018

plot_land_use_changes(year_from = 1990, year_to = 2018, data_path= data_path)

}
