% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_profiles.R
\name{plot_profiles}
\alias{plot_profiles}
\title{Plot CTD Profiles}
\usage{
plot_profiles(
  data,
  width,
  alpha,
  iterations = 1,
  type = c("Chlorophyll", "Turbidity"),
  min.depth = 0,
  directory = getwd(),
  plots = "any"
)
}
\arguments{
\item{data}{Data frame containing data from .vp2 files as obtained from
"read_vp2()".}

\item{width}{Window width against which each data-point is compared along the
CTD depth profile. Units should match those expressed in the "depth" column
of the input data.}

\item{alpha}{Level of significance used to test the incongruence of
data-points. Larger alpha-values make the function more sensitive in the
identification of incongruents.}

\item{iterations}{Number of iterations for applying the incongruence labeling
test. Points labeled "incongruent" are ignored in subsequent iterations.}

\item{type}{A character string. Either "Chlorophyll" or "Turbidity" for the
appropriate CTD measurement type.}

\item{min.depth}{Minimum depth above which points are excluded from the
profiles. Units should match those expressed in the "depth" column of the
input data.}

\item{directory}{The directory in which to save the "profiles.pdf" file. The
working directory is used by default.}

\item{plots}{Vector of character strings indicating which variables to use as
criteria for selecting profiles to plot in the final pdf. All the profiles
from a cast will be plotted if an incongruent data point was labeled for the
selected variables. The default "any" only excludes plots from casts that
had no incongruent data points for all variables. The option "all" plots
all casts regardless of whether they had incongruent data points.}
}
\value{
Data frame containing the input data-frame with columns labeling the
incongruence of each point for each variable; the "warning" column indicating
whether a certain measurement depth has incongruent values for more than two
variables. A "profiles" pdf is also created in the working directory
including plotted profiles for which Chlorophyll or Turbidity points were
identified as incongruent. Red points indicate those labeled as incongruent.
}
\description{
Plots CTD profiles for all variables measured by the CTD are included. Data
identified as incongruent are shown as red points, otherwise they are black.
The blue line is computed as the moving average with a window of 5 data
points, ignoring data identified as incongruent.
}
\examples{
## Load in example data and set it up in the temporary directory:
setup_example()

## Read the example casts:
casts <- read_vp2(directory = tempdir(), type = "Chlorophyll", ID = 12345)

## Label incongruents in the imported data and create a pdf with appropriate
## plots:
output <- plot_profiles(
  data = casts, width = 1, alpha = 0.0001,
  type = "Chlorophyll", min.depth = 1.5,
  directory = tempdir()
)

## To delete the example pdf you can run the following:
unlink(paste0(tempdir(), "/profiles.pdf"))

## To clear the temporary directory after using 'setup_example()':
unlink(paste0(tempdir(), "\\\\", list.files(tempdir(), pattern = ".vp2")))

}
