% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ign_data.R
\name{retrieve_cities}
\alias{retrieve_cities}
\title{Save the data of the IGN (french) cities classification.
The data is stored in the "data -> ign" directory}
\usage{
retrieve_cities(path_cities, path_cities_category)
}
\value{
None
}
\description{
Save the data of the IGN (french) cities classification.
The data is stored in the "data -> ign" directory
}
