% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ign_data.R
\name{retrieve_cities_ch}
\alias{retrieve_cities_ch}
\title{Save the data of the swiss cities classification.
The data is stored in the "data -> arep" directory}
\usage{
retrieve_cities_ch(path_cities_ch)
}
\value{
None
}
\description{
Save the data of the swiss cities classification.
The data is stored in the "data -> arep" directory
}
