% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ign_data.R
\name{retrieve_cities_fr}
\alias{retrieve_cities_fr}
\title{Save the data of the IGN (french) cities classification.
The data is stored in the "data -> arep" directory}
\usage{
retrieve_cities_fr(path_cities_fr, path_cities_category_fr)
}
\value{
None
}
\description{
Save the data of the IGN (french) cities classification.
The data is stored in the "data -> arep" directory
}
