% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ign_data.R
\name{retrieve_departements_fr}
\alias{retrieve_departements_fr}
\title{Save the data of french departments classification.
The data is stored in the "data -> arep" directory}
\usage{
retrieve_departements_fr(path_departments_fr)
}
\value{
None
}
\description{
Save the data of french departments classification.
The data is stored in the "data -> arep" directory
}
