% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ign_data.R
\name{retrieve_regions_fr}
\alias{retrieve_regions_fr}
\title{Save the data of french regions classification.
The data is stored in the "data -> arep" directory}
\usage{
retrieve_regions_fr(path_regions_fr)
}
\value{
None
}
\description{
Save the data of french regions classification.
The data is stored in the "data -> arep" directory
}
