% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_ign_data.R
\name{retrive_cities_ch}
\alias{retrive_cities_ch}
\title{Save the data of the swiss cities classification.
The data is stored in the "data -> ign" directory}
\usage{
retrive_cities_ch(path_cities_ch)
}
\value{
None
}
\description{
Save the data of the swiss cities classification.
The data is stored in the "data -> ign" directory
}
