% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incongruence_test.R
\name{run_inc_test}
\alias{run_inc_test}
\title{Run Incongruence Test}
\usage{
run_inc_test(column, iterations, df1, W, alpha, method)
}
\arguments{
\item{column}{Name of the variable being labeled for incongruents in the
input data-frame "df1".}

\item{iterations}{Number of iterations for applying the incongruence labeling
test. Points labeled "incongruent" are ignored in subsequent iterations.}

\item{df1}{Data frame containing data from .vp2 files as obtained from
\code{\link[=read_vp2]{read_vp2()}}.}

\item{W}{Halved window-length; inherited from \code{\link[=label_incongruents]{label_incongruents()}}.}

\item{alpha}{Level of significance used to test the incongruence of
data-points. Larger alpha-values make the function more sensitive in the
identification of incongruents.}

\item{method}{Method to be used to determine whether a point is an outlier.
Should be one of "t.student", "max.residual", "chisq", "dixon", or "grubbs".}
}
\value{
Data frame containing the original input data, followed by columns
indicating the incongruent labeling and p-value outcomes of each data point
for each variable.
}
\description{
Applies \code{\link[=incongruence_test]{incongruence_test()}} to each cast in a specified column from
the input data frame. This is done for the specified number of iterations.
}
