% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_territory.R
\name{select_territory}
\alias{select_territory}
\title{Crop the territory perimeter as one multipolygon
The function returns the 'box' of the created perimeter
and store the gpkg file in 'data->arep' folder}
\usage{
select_territory(
  regions_fr = NULL,
  departments_fr = NULL,
  epcis_fr = NULL,
  communes_ch = NULL,
  communes_fr = NULL
)
}
\value{
the coordinates of the box that delineates the territory
}
\description{
Crop the territory perimeter as one multipolygon
The function returns the 'box' of the created perimeter
and store the gpkg file in 'data->arep' folder
}
