% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vp2.R
\name{setup_example}
\alias{setup_example}
\title{Setup Example CTD Files}
\usage{
setup_example()
}
\description{
Sets up the files from ctdat.RData in a temporary directory to read with
\code{\link[=read_vp2]{read_vp2()}}.
}
\examples{
setup_example() # Saves example CTD files in a temporary directory

## These can be read as:
casts <- read_vp2(directory = tempdir(), type = "Chlorophyll", ID = 12345)

## To clear the temporary directory after using 'setup_example()':
unlink(paste0(tempdir(), "/", list.files(tempdir(), pattern = ".vp2")))

}
