% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SILViA-package.R
\docType{package}
\name{SILViA-package}
\alias{SILViA}
\alias{SILViA-package}
\title{SILViA: Sample Incongruents Labeling and Visualization Approach}
\description{
Method to identify outlying points recorded by a Valeport (https://www.valeport.co.uk/) Conductivity, Temperature, and Depth (CTD) meter along a depth profile, which are not congruent with the range of measurements taken in the surrounding water column. These are labeled as incongruent points. For this, the package includes a function to import and format .vp2 files, as well as functions to visualize depth profiles and the incongruent points they contain.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/KemenesT/SILViA}
  \item Report bugs at \url{https://github.com/KemenesT/SILViA/issues}
}

}
\author{
\strong{Maintainer}: Thomas Kemenes van Uden \email{thomaskemenes@gmail.com} (\href{https://orcid.org/0000-0002-0331-8538}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Erik Meesters (\href{https://orcid.org/0000-0002-3877-1164}{ORCID})
}

}
\keyword{internal}
