% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_apicarto_commune.R
\name{test_apicarto_commune}
\alias{test_apicarto_commune}
\title{Test apicarto_commune function
Check if the insee code exists}
\usage{
test_apicarto_commune(city_num)
}
\arguments{
\item{city_num}{the insee code of the city}
}
\value{
either an error message, or nothing
}
\description{
Test apicarto_commune function
Check if the insee code exists
}
