% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_old_epcis.R
\name{update_epcis}
\alias{update_epcis}
\title{Update old epcis found in aldo to the new classification}
\usage{
update_epcis(dt, col_name)
}
\arguments{
\item{dt}{a data table with old epci codes}

\item{col_name}{name of the column with epci code}
}
\value{
dt a data table with updated epci codes
}
\description{
Update old epcis found in aldo to the new classification
}
