% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_old_epcis.R
\name{update_outdated_epcis}
\alias{update_outdated_epcis}
\title{Update outdated epcis}
\usage{
update_outdated_epcis()
}
\value{
None
}
\description{
Write a file that matches old epcis found in aldo with the new classification
}
\details{
The function write a \code{epcis_old_new.csv} file with the corresponding old/new epcis.
}
