% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp2_as_oce.R
\name{vp2.as.oce}
\alias{vp2.as.oce}
\title{Convert Cast Data to oce Object}
\usage{
vp2.as.oce(data, file, directory, type = c("Chlorophyll", "Turbidity"))
}
\arguments{
\item{data}{A data-frame with .vp2 CTD data as returned by
\code{\link[=read_vp2]{read_vp2()}}.}

\item{file}{Filename of the cast to be converted into an oce object.}

\item{directory}{Directory where the raw .vp2 file corresponding to the cast
to be converted is stored. This is necessary to extract the meta-data that
is included for the file in the oce object.}

\item{type}{A character string. Either "Chlorophyll" or "Turbidity" for the
appropriate CTD measurement type.}
}
\value{
An oce object of class "ctd" containing the data for a single .vp2
file.
}
\description{
For a data-frame as obtained by \code{\link[=read_vp2]{read_vp2()}}, converts a single cast
file into an oce object applying \link[oce:as.ctd]{oce::as.ctd}.
}
