% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_old_epcis.R
\name{write_update_epcis}
\alias{write_update_epcis}
\title{Update old epcis found in aldo to the new classification}
\usage{
write_update_epcis()
}
\value{
None
}
\description{
Update old epcis found in aldo to the new classification
}
