% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_aldo_data.R
\name{carbon_content}
\alias{carbon_content}
\title{Extracts the data for the amount of carbon stored in forests from the Aldo
Excel tool.}
\usage{
carbon_content(path_to_aldo)
}
\arguments{
\item{path_to_aldo}{}
}
\value{
A data.table object.
}
\description{
Extracts the data for the amount of carbon stored in forests from the Aldo
Excel tool.
}
