% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gpkg_files.R
\name{download_gpkg_files}
\alias{download_gpkg_files}
\title{Download the territory chosen with the data from Corine Land Cover
of the chosen years, using 'happign' package
The function returns nothing but store the gpkg file
in the "data->arep" folder}
\usage{
download_gpkg_files(years, bd_foret = F)
}
\value{
None
}
\description{
Download the territory chosen with the data from Corine Land Cover
of the chosen years, using 'happign' package
The function returns nothing but store the gpkg file
in the "data->arep" folder
}
