% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_gpkg_files.R
\name{download_impermability_layers}
\alias{download_impermability_layers}
\title{Download impermability layers on the territory, for the selected years,
using 'happign' package
The function returns nothing but store raster files
in the "data->copernicus" folder}
\usage{
download_impermability_layers(shape, years = c("12", "15"))
}
\value{
None
}
\description{
Download impermability layers on the territory, for the selected years,
using 'happign' package
The function returns nothing but store raster files
in the "data->copernicus" folder
}
