% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_territory_files.R
\name{download_territory_files}
\alias{download_territory_files}
\title{Download territory files}
\usage{
download_territory_files(
  communes_fr = list(),
  epcis_fr = list(),
  departments_fr = list(),
  regions_fr = list(),
  years,
  data_path
)
}
\arguments{
\item{years}{years to download (1990, 2000, 2006, 2012, 2018)}

\item{data_path}{path to where the data is stored}

\item{bd_foret}{boolean to download bd_foret}
}
\description{
\code{download_territory_files} downloads all the necessary files related to the territory.

communes, EPCI, départements, and régions INSEE codes are  referenced here
\href{https://www.insee.fr/fr/information/6051727}{INSEE's codes}
}
\examples{
# Download files for the "Grand Annecy" (EPCI) territory, for all available years

download_territory_files(epcis_fr = "200066793", years= c(1990, 2000, 2006, 2012, 2018), data_path = data_path)


}
