% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_apigouv_insee_codes.R
\name{get_apigouv_insee_codes}
\alias{get_apigouv_insee_codes}
\title{Apigouv insee codes}
\usage{
get_apigouv_insee_codes(
    epcis_fr,
    regions_fr)
}
\arguments{
\item{communes_fr}{A \code{character} or a list of \code{character} of french
cities insee codes}

\item{epcis_fr}{A \code{character} or a list of \code{character} of french
epcis insee codes}

\item{departments_fr}{A \code{character} or a list of \code{character} of french
department insee codes}

\item{regions_fr}{A \code{character} or a list of \code{character} of french
regions insee codes}
}
\value{
\code{get_apigouv_insee_codes} returns a list of cities and departments insee codes
}
\description{
Retrieve insee codes from either epcis codes of departement codes
Codes of the different french territories
(communes, EPCI, départements, régions) are  referenced here
\href{https://www.insee.fr/fr/information/6051727}{INSEE's codes}
}
\details{
The function returns the map of the created perimeter
and store the gpkg file in 'data->arep' folder as "territory.gpkg".
}
\examples{

# Retrieve insee city codes in the epci of "Grand Annecy"
get_apigouv_insee_codes(epcis_fr = "200066793")

select_territory(regions_fr = "93", departments_fr = "31", epcis_fr = c("246900740", "200046977"), communes_fr = "81004")
}
