% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_carbon_stocks_balance.R
\name{get_carbon_stocks_balance}
\alias{get_carbon_stocks_balance}
\title{Compute total carbon stocks at a certain date, of a chosen region.
Four flows categories : stocks in biomass in forest, stocks in biomass out forest,
stocks in soils, and total cabon stocks}
\usage{
get_carbon_stocks_balance(stocks)
}
\arguments{
\item{stocks}{st object returned by 'get_carbon_storage' funciton}
}
\value{
a data.table with the flows in ktCO2/year
}
\description{
Compute total carbon stocks at a certain date, of a chosen region.
Four flows categories : stocks in biomass in forest, stocks in biomass out forest,
stocks in soils, and total cabon stocks
}
