% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forest_flows.R
\name{get_forest_flows}
\alias{get_forest_flows}
\title{Get forest flows}
\usage{
get_forest_flows(data_path)
}
\arguments{
\item{data_path}{path to where the data is stored}
}
\value{
forest flows in tC.an-1
}
\description{
Retrieve the carbon flows in the forests,
using the BD foret 2
}
\examples{

# flows <- get_carbon_flows(year_from = 1990, year_to = 2018, data_path = data_path)

}
