% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_land_use_changes.R
\name{get_land_use_changes}
\alias{get_land_use_changes}
\title{Get land use changes}
\usage{
get_land_use_changes(year_from, year_to, data_path)
}
\arguments{
\item{year_from}{the CLC reference year (1990, 2000, 2006, 2012, 2018)}

\item{year_to}{the year to be compared to the reference year,
must be superior to reference year}

\item{data_path}{path to where the data is stored}
}
\value{
a sf object with the land use changes between year_to and year_from
}
\description{
\code{get_land_use_changes} collects land use changes between two years
}
