% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/land_use_changes.R
\name{get_land_use_changes}
\alias{get_land_use_changes}
\title{Collect land use changes between two years}
\usage{
get_land_use_changes(year_from, year_to, remove_unchanged = TRUE)
}
\value{
a sf object with the land use changes between year_to and year_from
}
\description{
Collect land use changes between two years
}
