% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incongruence_test.R
\name{incongruence_test}
\alias{incongruence_test}
\title{Incongruence Test}
\usage{
incongruence_test(point_time, column, subdf, W, alpha, casts, cast)
}
\arguments{
\item{point_time}{Time measurement of the data point to be tested for
incongruence.}

\item{column}{Name of the variable being labeled for incongruents in the
input data-frame "df1".}

\item{subdf}{Subset data.frame inherited from "run_inc_test".}

\item{W}{Halved window-length; inherited from "LabelIncongruents".}

\item{alpha}{Level of significance used to test the incongruence of
data-points. Larger alpha-values make the function more sensitive in the
identification of incongruents.}

\item{casts}{Vector of all cast names in the data-frame being processed;
inherited from "run_inc_test".}

\item{cast}{Name of the active cast; inherited from "run_inc_test".}
}
\value{
Data frame containing the T-score and incongruence labeling for each
data point in the active cast.
}
\description{
Tests each point in the input vector for incongruence with the other elements
within the specified depth window.
}
