% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ign_data.R
\name{load_ign_aldo_data}
\alias{load_ign_aldo_data}
\title{download the files needed to run the tool}
\usage{
load_ign_aldo_data()
}
\value{
None
}
\description{
download the files needed to run the tool
}
