% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flows.R
\name{map_carbon_flows}
\alias{map_carbon_flows}
\title{Plot carbon flows of a chosen region.}
\usage{
map_carbon_flows(flows)
}
\arguments{
\item{flows}{st object returned by 'get_carbon_flows' funciton}
}
\value{
a data.table with the flows in ktCO2/year
}
\description{
Plot carbon flows of a chosen region.
}
