% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carbon_stocks_mapping.R
\name{map_carbon_storage}
\alias{map_carbon_storage}
\title{Map the carbon storage on the chosen region.}
\usage{
map_carbon_storage(dt)
}
\value{
a ggplot with the map
}
\description{
Map the carbon storage on the chosen region.
}
