% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_and_plots.R
\name{map_forest_flows}
\alias{map_forest_flows}
\title{Plot carbon flows in forest of a chosen region.}
\usage{
map_forest_flows()
}
\value{
map of carbon flows by the forest
}
\description{
Plot carbon flows in forest of a chosen region.
}
