% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_and_plots.R
\name{plot_land_use_changes}
\alias{plot_land_use_changes}
\title{Plot land use changes between two year.
Results are presented in a sankey diagram}
\usage{
plot_land_use_changes(year_from, year_to, level = "soft")
}
\value{
The sankey diagram
}
\description{
Plot land use changes between two year.
Results are presented in a sankey diagram
}
