% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_and_plots.R
\name{plot_land_use_changes}
\alias{plot_land_use_changes}
\title{Plot land use changes between two year.
Results are presented in a sankey diagram}
\usage{
plot_land_use_changes(year_from, year_to, data_path, level = "soft")
}
\arguments{
\item{year_from}{the CLC reference year (1990, 2000, 2006, 2012, 2018)}

\item{year_to}{the year to be compared to the reference year,
must be superior to reference year}

\item{data_path}{path to where the data is stored}
}
\value{
The sankey diagram
}
\description{
Plot land use changes between two year.
Results are presented in a sankey diagram
}
