% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incongruence_test.R
\name{run_inc_test}
\alias{run_inc_test}
\title{Run Incongruence Test}
\usage{
run_inc_test(column, iterations, df1, W, alpha)
}
\arguments{
\item{column}{Name of the variable being labeled for incongruents in the
input data-frame "df1".}

\item{iterations}{Number of iterations for applying the incongruence labeling
test. Points labeled "incongruent" are ignored in subsequent iterations.}

\item{df1}{Data frame containing data from .vp2 files as obtained from
"read_vp2()".}

\item{W}{Halved window-length; inherited from "LabelIncongruents".}

\item{alpha}{Level of significance used to test the incongruence of
data-points. Larger alpha-values make the function more sensitive in the
identification of incongruents.}
}
\value{
Data frame containing the original input data, followed by columns
indicating the incongruent labeling and T-score outcomes of each data point
for each variable.
}
\description{
Applies incongruence_test() to each cast in a specified column from the input
data frame. This is done for the specified number of iterations.
}
