% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_territory.R
\name{select_territory}
\alias{select_territory}
\title{Select territory}
\usage{
select_territory(
    regions_fr = list(),
    departments_fr = list(),
    epcis_fr= list(),
    communes_fr = list(),
    communes_ch= list(),
    local = F)
}
\arguments{
\item{regions_fr}{A \code{character} or a list of \code{character} of french
region codes to be included in the perimeter}

\item{departments_fr}{A \code{character} or a list of \code{character} of french
departements codes to be included in the perimeter}

\item{epcis_fr}{A \code{character} or a list of \code{character} of french
EPCIs codes to be included in the perimeter}

\item{communes_fr}{A \code{character} or a list of \code{character} of french
municipalities codes to be included in the perimeter}

\item{local}{If TRUE, the data is loaded from the "arep" folder. If false, a request is done on a api.}
}
\value{
Map of the selected territory
}
\description{
Crop spatial boundaries to be studied.
Codes of the different french territories
(communes, EPCI, départements, régions) are  referenced here
\href{https://www.insee.fr/fr/information/6051727}{INSEE's codes}
}
\details{
The function returns the map of the created perimeter
and store the gpkg file in 'data->arep' folder as "territory.gpkg".
}
\examples{
# Shape of the "Grand Annecy" EPCI

select_territory(epcis_fr = "200066793")


# Shape of the union of a region (PACA 93), a departement (Occitanie 31),
two epcis (in the Rhône department), and one city (Albi 81004)

select_territory(regions_fr = "93", departments_fr = "31", epcis_fr = c("246900740", "200046977"), communes_fr = "81004")
}
