% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_territory.R
\name{select_territory}
\alias{select_territory}
\title{Select territory}
\usage{
select_territory(communes_fr, epcis_fr, departments_fr, regions_fr,data_path)
}
\arguments{
\item{communes_fr}{A \code{character} or a list of \code{character} of french
municipalities codes to be included in the perimeter}

\item{epcis_fr}{A \code{character} or a list of \code{character} of french
EPCIs codes to be included in the perimeter}

\item{departments_fr}{A \code{character} or a list of \code{character} of french
departements codes to be included in the perimeter}

\item{regions_fr}{A \code{character} or a list of \code{character} of french
region codes to be included in the perimeter}

\item{data_path}{path to where the data is stored}
}
\value{
shape of the selected territory
}
\description{
\code{select_territory} crops spatial boundaries to be studied.
Codes of the different french territories
(communes, EPCI, départements, régions) are  referenced here
\href{https://www.insee.fr/fr/information/6051727}{INSEE's codes}

The function returns the map of the created perimeter
and store the gpkg file in 'data->arep' folder as "territory.gpkg".
}
