% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_path.R
\name{setup_path}
\alias{setup_path}
\title{Setup Path}
\usage{
setup_path(data_path)
}
\arguments{
\item{data_path}{A \code{character} or a list of \code{character} of french
region codes to be included in the perimeter}
}
\value{
data_path
}
\description{
\code{setup_path} ...
}
