% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vp2_as_oce.R
\name{vp2.as.oce}
\alias{vp2.as.oce}
\title{Convert vp2 Cast Data to oce Object}
\usage{
vp2.as.oce(data, file, directory)
}
\arguments{
\item{data}{A data-frame with .vp2 CTD data as returned by
\code{\link[=read_vp2]{read_vp2()}}.}

\item{file}{Filename of the cast to be converted into an oce object.}

\item{directory}{Directory where the raw .vp2 file corresponding to the cast
to be converted is stored. This is necessary to extract the
meta-data included included in the oce object.}
}
\value{
An oce object of class "ctd" containing the data for a single .vp2
file.
}
\description{
For a data-frame as obtained by \code{\link[=read_vp2]{read_vp2()}}, converts a single cast
file into an oce object.
}
\examples{
## Load in example data and set it up in the temporary directory:
setup_example()

## Read the example casts:
casts <- read_vp2(directory = tempdir(), ID = 12345)

## Subset single cast:
file_name <- unique(casts$filename)[1]
cast <- casts[which(casts$filename == file_name), ]

## Convert single cast to oce object:
object <- vp2.as.oce(data = cast, file = file_name, directory = tempdir())

## To clear the temporary directory after using 'setup_example()':
unlink(paste0(tempdir(), "/", list.files(tempdir(), pattern = ".vp2")))

}
