% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_region.R
\name{write_gpkg_files}
\alias{write_gpkg_files}
\title{Merge the territory chosen with the data from Corine Land Cover
of the chosen years, using local files
The function returns nothing but store the gpkg file
in the "data->arep" folder}
\usage{
write_gpkg_files(box, years)
}
\value{
None
}
\description{
Merge the territory chosen with the data from Corine Land Cover
of the chosen years, using local files
The function returns nothing but store the gpkg file
in the "data->arep" folder
}
